% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{corrImg}
\alias{corrImg}
\title{Visualizing genomic correlation}
\usage{
corrImg(corrs, ...)
}
\arguments{
\item{corrs}{a numeric correlation matrix}

\item{...}{optional arguments to pass to `image`}
}
\value{
Returns nothing, but plots the correlations in a heat map
in the active device.
}
\description{
Visualizing genomic correlation
}
\details{
A simple wrapper for `image` that places the correlation
in a more intuitive form where the 1,1 entry is in the top left.
}
\author{
Chris Salahub
}
