% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeGenerators.R
\name{generateLocations}
\alias{generateLocations}
\title{Generate locations for markers}
\usage{
generateLocations(nmark, locFuns, ...)
}
\arguments{
\item{nmark}{vector of integers giving the number of locations by
chromosome}

\item{locFuns}{list of functions which each accept an integer
and return a numeric vector with length equal to that integer,
recycled if not of the same length as `nmark`}

\item{...}{list of additional arguments to each function in
`locFuns`, recycled if not the same length}
}
\value{
A list of numeric vectors giving the locations of markers
on each chromosome.
}
\description{
Generate locations for markers
}
\details{
A wrapper function for location-generating functions that
accepts a list of functions and an integer vector of counts of the
same length and returns a list of locations resulting from the
application of the functions to the corresponding counts.
}
\author{
Chris Salahub
}
