% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{popCorrelation}
\alias{popCorrelation}
\title{Compute observed correlation for a `population`}
\usage{
popCorrelation(population, scoring = scoreAdditive, ...)
}
\arguments{
\item{population}{`population` object}

\item{scoring}{function which accepts an encoding matrix and
returns a numeric vector with one numeric for each row of the
encoding matrix}

\item{...}{additional arguments to pass to `cor`}
}
\value{
A correlation matrix between markers in `population`.
}
\description{
Compute observed correlation for a `population`
}
\details{
A wrapper which computes the correlation for a
`population` given a scoring function to apply to each
encoding.
}
\author{
Chris Salahub
}
