% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{subsetPopulation}
\alias{subsetPopulation}
\title{Subsetting populations by markers}
\usage{
subsetPopulation(
  population,
  markInd = (1:length(population$chromosome)),
  genInd = (1:length(population$encoding))
)
}
\arguments{
\item{population}{`population` to be subset by markers}

\item{markInd}{indices of markers to keep}

\item{genInd}{indices of individuals to keep}
}
\value{
A population object including only the encoding matrices
of `genInd` subset to the markers of `markInd`.
}
\description{
Subsetting populations by markers
}
\details{
Like the `genome` analog, this subsets markers, drops
unused chromosomes, and returns a new population object. Unlike
the single `genome` case, this includes a second argument to
drop individual genomes from the population.
}
\author{
Chris Salahub
}
